mysql_backend_clean/requirements.txt
# 經濟腦TA 後端服務依賴套件
fastapi>=0.95.0
uvicorn>=0.21.1
pymysql>=1.0.3
pandas>=1.5.0
openpyxl>=3.1.0
python-multipart>=0.0.6
python-dotenv>=0.21.0
pydantic>=1.10.7
